add return value check

+++ imgcpy.sh
@@ -13,6 +13,8 @@ loop_flag=1
 while [ "${loop_flag}" != "0" ]
 do
         sd_path=$(ls --sort=time /dev/sd* | awk '{print $1}' | sed -n '1p')
+        test ! $? -eq 0 && echo "[ERROR] detecting device failed" && exit 0
+
         test ! -e ${sd_path} && echo "[ERROR] "$sd_path" does not exit." && exit 0
 
         echo "Detect the newest device name = "$sd_path
@@ -38,6 +40,7 @@ do
 done
 
 sd_mount_point=$(mount -l | grep ${sd_path} | awk '{print $3}' | sed -n '1p')
+test ! $? -eq 0 && echo "[ERROR] detect mount point failed" && exit 0
 
 # if your os doesn't mount it for you automatically...
 # then create a new folder ~/sdcard_tmp123456789 and mount it.
@@ -49,15 +52,20 @@ if [ ! ${sd_mount_point} ]; then
         test ! -d ${sd_mount_point} && mkdir ${sd_mount_point}
 
         sudo mount ${sd_path} ${sd_mount_point}
+        test ! $? -eq 0 && echo "[ERROR] mount failed" && exit 0
         echo "The mount point of SD card is "${sd_mount_point}" (mounted by the script automatically)"
         
+
         sudo cp $PWD/build/kernel8.img ${sd_mount_point}
+        test ! $? -eq 0 && echo "[ERROR] cp failed" && exit 0
 
         sync
         echo "Finish copying Kernel8.img to "$sd_path
         sudo umount ${sd_mount_point}
-        
+        test ! $? -eq 0 && echo "[ERROR] umount failed" && exit 0
+
         sudo rm -r ${sd_mount_point}
+        test ! $? -eq 0 && echo "[ERROR] rm failed" && exit 0
         
         exit 0
 fi
@@ -66,6 +74,7 @@ fi
 echo "The mount point of SD card is "${sd_mount_point}" (mounted by your system automatically)"
 
 cp $PWD/build/kernel8.img ${sd_mount_point}
+test ! $? -eq 0 && echo "[ERROR] mount failed" && exit 0
 
 sync
